Move business log out of handler * Remove  unnecessary test base class
